applied some test refactorings and cs fixes
