+++ tests/src/CL/Slack/Tests/Payload/UsersGetPresencePayloadResponseTest.php
