+++ tests/src/CL/Slack/Tests/Payload/UsersGetPresencePayloadTest.php
